requirements.txt
certifi==2024.8.30
charset-normalizer==3.3.2
colorama==0.4.6
filelock==3.15.4
fsspec==2024.9.0
idna==3.8
Jinja2==3.1.4
llvmlite==0.43.0
MarkupSafe==2.1.5
more-itertools==10.4.0
mpmath==1.3.0
networkx==3.3
numba==0.60.0
numpy==2.0.2
openai-whisper==20231117
pydub==0.25.1
regex==2024.7.24
requests==2.32.3
sympy==1.13.2
tiktoken==0.7.0
torch==2.4.1
tqdm==4.66.5
typing_extensions==4.12.2
urllib3==2.2.2